+++ gql_websocket_link/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+The custom client has been removed.
+The library is now using WebSocketChannel as a base for the connection. 
+
 ## 0.1.0
 
 - Websocket link